feat: block all replciation and query rpc calls until wait for grants has succeeded
