+++ go/vt/vttablet/tabletmanager/rpc_replication_test.go
@@ -1,0 +1,44 @@
+/*
+Copyright 2023 The Vitess Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package tabletmanager
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+// TestWaitForGrantsToHaveApplied tests that waitForGrantsToHaveApplied only succeeds after waitForDBAGrants has been called.
+func TestWaitForGrantsToHaveApplied(t *testing.T) {
+	tm := &TabletManager{
+		_waitForGrantsComplete: make(chan struct{}),
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
+	defer cancel()
+	err := tm.waitForGrantsToHaveApplied(ctx)
+	require.ErrorContains(t, err, "deadline exceeded")
+
+	err = tm.waitForDBAGrants(nil, 0)
+	require.NoError(t, err)
+
+	secondContext, secondCancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
+	defer secondCancel()
+	err = tm.waitForGrantsToHaveApplied(secondContext)
+	require.NoError(t, err)
+}
