+++ go/vt/vttablet/tabletserver/tabletserver.go
@@ -43,7 +43,6 @@ import (
 	"vitess.io/vitess/go/trace"
 	"vitess.io/vitess/go/vt/callerid"
 	"vitess.io/vitess/go/vt/dbconfigs"
-	"vitess.io/vitess/go/vt/dbconnpool"
 	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/logutil"
 	"vitess.io/vitess/go/vt/mysqlctl"
@@ -240,41 +239,6 @@ func NewTabletServer(ctx context.Context, name string, config *tabletenv.TabletC
 	tsv.registerDebugEnvHandler()
 
 	return tsv
-}
-
-// WaitForDBAGrants waits for DBA user to have the required privileges to function properly.
-func WaitForDBAGrants(config *tabletenv.TabletConfig, waitTime time.Duration) error {
-	// We don't wait for grants if the tablet is externally managed. Permissions
-	// are then the responsibility of the DBA.
-	if config == nil || config.DB.HasGlobalSettings() || waitTime == 0 {
-		return nil
-	}
-	timer := time.NewTimer(waitTime)
-	ctx, cancel := context.WithTimeout(context.Background(), waitTime)
-	defer cancel()
-	for {
-		conn, err := dbconnpool.NewDBConnection(ctx, config.DB.DbaConnector())
-		if err == nil {
-			res, fetchErr := conn.ExecuteFetch("SHOW GRANTS", 1000, false)
-			if fetchErr != nil {
-				log.Errorf("Error running SHOW GRANTS - %v", fetchErr)
-			}
-			if fetchErr == nil && res != nil && len(res.Rows) > 0 && len(res.Rows[0]) > 0 {
-				privileges := res.Rows[0][0].ToString()
-				// In MySQL 8.0, all the privileges are listed out explicitly, so we can search for SUPER in the output.
-				// In MySQL 5.7, all the privileges are not listed explicitly, instead ALL PRIVILEGES is written, so we search for that too.
-				if strings.Contains(privileges, "SUPER") || strings.Contains(privileges, "ALL PRIVILEGES") {
-					return nil
-				}
-			}
-		}
-		select {
-		case <-timer.C:
-			return fmt.Errorf("waited %v for dba user to have the required permissions", waitTime)
-		default:
-			time.Sleep(100 * time.Millisecond)
-		}
-	}
 }
 
 func (tsv *TabletServer) loadQueryTimeout() time.Duration {
